% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netSEM-package.R
\docType{package}
\name{netSEM-package}
\alias{netSEM}
\alias{netSEM-package}
\title{netSEM: Network Structural Equation Modeling}
\description{
The network structural equation modeling conducts a network statistical analysis on a data frame of coincident observations of multiple continuous variables [1]. It builds a pathway model by exploring a pool of domain knowledge guided candidate statistical relationships between each of the variable pairs, selecting the 'best fit' on the basis of a specific criteria such as adjusted r-squared value. This material is based upon work supported by the U.S. National Science Foundation Award EEC-2052776 and EEC-2052662 for the MDS-Rely IUCRC Center, under the NSF Solicitation: NSF 20-570 Industry-University Cooperative Research Centers Program [1] Bruckman, Laura S., Nicholas R. Wheeler, Junheng Ma, Ethan Wang, Carl K. Wang, Ivan Chou, Jiayang Sun, and Roger H. French. (2013) \doi{10.1109/ACCESS.2013.2267611}.

The R package 'netSEM' conducts a net-SEM statistical analysis (network structural equation modeling) on a dataframe of coincident observations of multiple continuous variables.
}
\references{
1. Bruckman, Laura S., Nicholas R. Wheeler, Junheng Ma, Ethan Wang, Carl K. Wang, Ivan Chou, Jiayang Sun, and Roger H. French. "Statistical and Domain Analytics Applied to PV Module Lifetime and Degradation Science." IEEE Access 1 (2013): 384-403. doi:10.1109/ACCESS.2013.2267611

2. Bruckman, Laura S., Nicholas R. Wheeler, Ian V. Kidd, Jiayang Sun, and Roger H. French. "Photovoltaic Lifetime and Degradation Science Statistical Pathway Development: Acrylic Degradation." In SPIE Solar Energy+ Technology, 8825:88250D-8. International Society for Optics and Photonics, 2013. doi:10.1117/12.2024717
}
\author{
\strong{Maintainer}: Laura S. Bruckman \email{lsh41@case.edu} (\href{https://orcid.org/0000-0003-1271-1072}{ORCID})

Authors:
\itemize{
  \item Wei-Heng Huang \email{wxh272@case.edu} (\href{https://orcid.org/0000-0002-6609-4981}{ORCID})
  \item Nicholas R. Wheeler \email{nrw16@case.edu} (\href{https://orcid.org/0000-0003-2248-8919}{ORCID})
  \item Addison G. Klinke \email{agk38@case.edu} (\href{https://orcid.org/0000-0002-6985-7657}{ORCID})
  \item Yifan Xu \email{yxx218@case.edu} (\href{https://orcid.org/0000-0003-1696-0228}{ORCID})
  \item Wenyu Du \email{wxd97@case.edu} (\href{https://orcid.org/0000-0002-8672-9104}{ORCID})
  \item Amit K. Verma \email{aak126@case.edu} (\href{https://orcid.org/0000-0003-1580-3001}{ORCID})
  \item Abdulkerim Gok \email{axg515@case.edu} (\href{https://orcid.org/0000-0003-3433-7106}{ORCID})
  \item Roger H. French \email{rxf131@case.edu} (\href{https://orcid.org/0000-0002-6162-0532}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Devin A. Gordon \email{dag109@case.edu} (\href{https://orcid.org/0000-0002-5919-0422}{ORCID}) [contributor]
  \item Yu Wang \email{yxw880@case.edu} (\href{https://orcid.org/0000-0003-1353-2578}{ORCID}) [contributor]
  \item Sameera Nalin Venkat \email{sxn440@case.edu} (\href{https://orcid.org/0000-0002-9114-4602}{ORCID}) [contributor]
  \item HeinHtet Aung \email{hxh483@case.edu} (\href{https://orcid.org/0000-0002-8191-8213}{ORCID}) [contributor]
  \item Yeajin Jo \email{yxj414@case.edu} (\href{https://orcid.org/0000-0001-7026-3779}{ORCID}) [contributor]
  \item Xuanji Yu \email{xxy530@case.edu} (\href{https://orcid.org/0000-0003-1967-3769}{ORCID}) [contributor]
  \item Kemal Ozdemirli \email{kxo146@case.edu} [contributor]
  \item Jonathan Gordon \email{jgordon5@bowdoin.edu} (\href{https://orcid.org/0009-0007-5958-7386}{ORCID}) [contributor]
  \item Jayvic Cristian Jimenez \email{jxj440@case.edu} (\href{https://orcid.org/0000-0002-2342-5648}{ORCID}) [contributor]
  \item Jiqi Liu \email{jxl1763@case.edu} (\href{https://orcid.org/0000-0003-2016-4160}{ORCID}) [contributor]
  \item Alan J. Curran \email{ajc269@case.edu} (\href{https://orcid.org/0000-0002-4505-8359}{ORCID}) [contributor]
  \item Justin S. Fada \email{jsf77@case.edu} (\href{https://orcid.org/0000-0002-0029-5051}{ORCID}) [contributor]
  \item Xuan Ma \email{xxm115@case.edu} (\href{https://orcid.org/0000-0003-2361-2846}{ORCID}) [contributor]
  \item Jennifer L. Braid \email{jlb269@case.edu} (\href{https://orcid.org/0000-0002-0677-7756}{ORCID}) [contributor]
  \item Jennifer L. W. Carter \email{jwc137@case.edu} (\href{https://orcid.org/0000-0001-6702-729X}{ORCID}) [contributor]
}

}
\keyword{internal}
