% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayRMSE.r
\name{pathwayRMSE}
\alias{pathwayRMSE}
\title{Summary of pathway RMSE}
\usage{
pathwayRMSE(x, maxlen = 2, ...)
}
\arguments{
\item{x}{An object of class "netSEM", the returned list from netSEMm().}

\item{maxlen}{The maximum length of chosen mechanism.}

\item{...}{A S3 generic/method consistency.}
}
\value{
A data frame of result. A summary of RMSE results for all of pathways.
}
\description{
summarize root mean square error (RMSE) for direct and indrect pathway from netSEM result
}
\details{
pathwayRMSE gives a summary about RMSE
}
\examples{
data(acrylic)
ans <- netSEMm(acrylic)
pathwayRMSE(ans,maxlen=2)

}
\seealso{
\link[netSEM]{netSEMm}, \link[netSEM]{pathwayPredict}
}
