% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.netSEMp2.R
\name{summary.netSEMp2}
\alias{summary.netSEMp2}
\title{Summary of netSEMp2}
\usage{
\method{summary}{netSEMp2}(object, ...)
}
\arguments{
\item{object}{An object of class "netSEMp2", the returned list from netSEMp2().}

\item{...}{A S3 generic/method consistency.}
}
\value{
NULL. A summary of data and fitting result is printed on screen.
}
\description{
summarize netSEM result
}
\details{
summary.netSEMp2 gives a breif summary about the netSEM Principle 2 analysis.
}
\examples{
data(acrylic)
ans <- netSEMp2(acrylic)
summary(ans)


}
